画面アスペクトがint型になっていたのでfloatに変更、各警告に対応しておきました。 DirectInputにたぶん対応できた。それに伴って基底クラスInputおよびInputKeyboardのメンバ変数(入力デバイス)の扱いに若干の変更あり。 あと何か変更した気がするけど忘れちゃった☆ あ、中村参上！

+++ gameProgram/debugproc.cpp
@@ -120,7 +120,7 @@ void CDebugproc::Draw(void)
 {
 	if(m_drawFlg)
 	{
-		RECT rect = {0, 0, SCREEN_WIDTH, SCREEN_HEIGHT};
+		RECT rect = { 0, 0, static_cast<LONG>(SCREEN_WIDTH), static_cast<LONG>(SCREEN_HEIGHT) };
 
 		m_font->DrawText(NULL, m_strDebug, -1, &rect,
 						 DT_LEFT, D3DCOLOR_ARGB(0xff, 0xff, 0xff, 0xff));
